api: exposing init wallet mock provider function
